Add "snapshot" test class for MD with neighbor list
This tests the consistency of the Integrator. First we run the simulation
for a certain number of steps and store the coordinates in a file. The
test then runs the simulation with the same random seeds and number of
steps and checks to make sure the coordinates are the same as stored in
the file. If any changes cause the Integrator to behave differently
unexpectedly it should hopefully cause the test to fail.