Fix failing tests - IteratorTestAbstract
Add a copy of the iterate to atoms ("new ArrayList(nextAtom)") to mimic
the old behavior which did "new AtomsetArray(nextAtom)". This is necessary
since the iterate is sometimes just a reference to a single list in the
iterator which is modified by each call to next and reset. Since the
iterator is reset before assertEquals is called this causes null or
incorrect references to be shared. This is why mutable state is bad...